pyproject.toml
[project]
name = "hihu-3-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "pillow>=12.0.0",
]
